requirements.txt
# DEPENDENCIES: List of required Python packages
fastapi
uvicorn[standard]
jinja2
python-multipart
pandas
scikit-learn
requests
python-dotenv

google-genai
xgboost
itsdangerous
sqlalchemy
